change everything to not use int

+++ NavNodeLink.cs
@@ -1,13 +1,13 @@
 namespace AStarSquares
 {
     public struct NavNodeLink {
-        public NavNodeLink(INavNode linkedNavNode, int moveCost, int vertical) {
+        public NavNodeLink(INavNode linkedNavNode, float moveCost, float vertical) {
             LinkedNavNode = linkedNavNode;
             Distance = moveCost;
             Vertical = vertical;
         }
-        public int Vertical;
+        public float Vertical;
         public INavNode LinkedNavNode;
-        public int Distance;
+        public float Distance;
     }    
 }
