+++ NavPath.cs
@@ -7,18 +7,18 @@ namespace AStarSquares
 
         public NavPath(IList<PathNode> pathNodes) {
             PathNodes = pathNodes;
-            TotalCost = pathNodes.Aggregate(0, (total, next) => total += next.Cost );
+            TotalCost = pathNodes.Aggregate(0f, (total, next) => total += next.Cost );
         }
-        public int TotalCost;
+        public float TotalCost;
         public IList<PathNode> PathNodes;
 
         public struct PathNode {
-            public PathNode(INavNode navNode, int cost) {
+            public PathNode(INavNode navNode, float cost) {
                 NavNode = navNode;
                 Cost = cost;
             }
             public INavNode NavNode;
-            public int Cost;
+            public float Cost;
         }
     }
